priority queue after changes
